+++ packages/config-ui/src/choice-configuration/index.jsx
@@ -11,7 +11,6 @@ import ActionDelete from 'material-ui-icons/Delete';
 import IconButton from 'material-ui/IconButton';
 import Typography from 'material-ui/Typography';
 
-
 const EditableHtmlContainer = withStyles(theme => ({
   labelContainer: {
 
@@ -56,7 +55,7 @@ const Feedback = withStyles(theme => ({
   }
 });
 
-class RawChoiceConfiguration extends React.Component {
+export class RawChoiceConfiguration extends React.Component {
 
 
   _changeFn = (key) => (update) => {
@@ -67,7 +66,15 @@ class RawChoiceConfiguration extends React.Component {
   }
 
   onLabelChange = this._changeFn('label');
-  onValueChange = this._changeFn('value');
+
+  onValueChange = (event) => {
+    const { onChange, data } = this.props;
+    const value = event.target.value;
+    if (onChange) {
+      onChange({ ...data, value });
+    }
+  }
+
   onCheckedChange = (event) => {
     const checked = event.target.checked;
     const { data, onChange } = this.props;
@@ -105,20 +112,21 @@ class RawChoiceConfiguration extends React.Component {
 
   render() {
 
-    const { data, classes, mode, onDelete, defaultFeedback, index } = this.props;
+    const { data, classes, mode, onDelete, defaultFeedback, index, className } = this.props;
 
     const InputToggle = mode === 'checkbox' ? InputCheckbox : InputRadio;
+    const names = classNames(classes.choiceConfiguration, className);
     return (
-      <div className={classes.choiceConfiguration}>
+      <div className={names}>
         <div className={classes.topRow}>
-          {index && <Typography
+          {index > 0 && <Typography
             className={classes.index}
             type="title">{index}</Typography>}
           <InputToggle
             className={classes.toggle}
             onChange={this.onCheckedChange}
             label={'Correct'}
-            checked={data.checked} />
+            checked={!!data.checked} />
           <TextField
             label={'Value'}
             value={data.value}